:sparkles: :art: Codigo de encuesta mejorado, validacion de checkboxes correcta
