+++ src/components/Navbar/Navbar.js
@@ -1,14 +1,6 @@
 import React, {useContext, useState} from 'react';
 import { FaBars } from 'react-icons/fa';
-import {
-  Nav,
-  NavLink,
-  Bars,
-  MobileIcon,
-  NavMenu,
-  NavBtn,
-  NavBtnLink 
-} from './NavbarElements';
+import { Nav,NavLink, MobileIcon, NavMenu, NavBtn, NavBtnLink } from './NavbarElements';
 import './Navbar.css'
 import logo from "../../imagenes/logo_coronanalyst.jpeg"
 import {MyContext} from '../../contexts/MyContext'
